move and rename DTO package and rename repository
