+++ src/main/java/org/mae/twg/backend/controllers/AuthController.java
@@ -4,9 +4,9 @@ import io.swagger.v3.oas.annotations.Operation;
 import io.swagger.v3.oas.annotations.tags.Tag;
 import jakarta.validation.Valid;
 import lombok.RequiredArgsConstructor;
-import org.mae.twg.backend.models.DTO.JwtAuthenticationResponse;
-import org.mae.twg.backend.models.DTO.SignInRequest;
-import org.mae.twg.backend.models.DTO.SignUpRequest;
+import org.mae.twg.backend.dto.JwtAuthenticationResponse;
+import org.mae.twg.backend.dto.SignInRequest;
+import org.mae.twg.backend.dto.SignUpRequest;
 import org.mae.twg.backend.services.AuthenticationService;
 import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestBody;
